7. Cross-Origin Resource Sharing (CORS)
Due to security concerns, it's not always so easy to make a cross-origin AJAX request to fetch a particular set of data. Instead, you'll often be met with the dreaded No 'Access-Control-Allow-Origin' header is present on the requested resource response. Let's discuss what this header means, and how we can fix it.